+++ acr122u_reader/custom_components/acr122u/const.py
@@ -3,6 +3,6 @@ EVENT_CARD_PRESENT = "acr122u_card_present"
 EVENT_CARD_REMOVED = "acr122u_card_removed"
 SIGNAL_UPDATE = f"{DOMAIN}_update"
 DEVICE_IDENTIFIER = "acr122u_usb_reader"
-PLATFORMS = ["binary_sensor", "sensor"]
+PLATFORMS = ["binary_sensor", "event", "sensor"]
 
 EVENT_DEVICE_ACTIVITY = f"{DOMAIN}_event"